Moved IRegistration to api directory
