+++ src/server/commands/api/IRegistration.ts
